Update migrations ids to laravel 8

+++ database/migrations/2020_12_29_205300_add_column_establishment_id_to_organizational_units_table.php
@@ -16,7 +16,7 @@ class AddColumnEstablishmentIdToOrganizationalUnitsTable extends Migration
         Schema::table('organizational_units', function (Blueprint $table) {
             //
             $table->tinyInteger('level')->nullable()->after('name');;
-            $table->unsignedInteger('establishment_id')->nullable()->after('organizational_unit_id');;
+            $table->unsignedBigInteger('establishment_id')->nullable()->after('organizational_unit_id');;
             $table->foreign('establishment_id')->references('id')->on('establishments');
         });
     }
