+++ database/migrations/2022_10_20_171155_create_rem_users_table.php
@@ -17,7 +17,7 @@ class CreateRemUsersTable extends Migration
             $table->id();
             $table->foreignId('user_id')
                 ->constrained('users');
-            $table->unsignedInteger('establishment_id');
+            $table->unsignedBigInteger('establishment_id');
             $table->foreign('establishment_id')->references('id')->on('establishments');
             $table->timestamps();
             $table->softDeletes();
